Create Tab Navigation component which nests JSX custom Tab components
