+++ src/App.tsx
@@ -2,45 +2,24 @@ import React from 'react';
 // import logo from './logo.svg';
 import './App.css';
 import TabsNavigation from './components/TabsNavigation/TabsNavigation';
+import PocketCustomTab from './components/TabsNavigation/Tab';
 
 function App() {
-    return (
-        <div className="App">
-            {/* <!-- your component goes here to test the rendering! --> */}
-            <TabsNavigation>
-                {[
-                    {
-                        NameTab: 'Prospects',
-                        BadgeTab: 99,
-                    },
-                    {
-                        NameTab: 'Snoozed',
-                        BadgeTab: 0,
-                    },
-                    {
-                        NameTab: 'Approved',
-                        BadgeTab: 5,
-                    },
-                    {
-                        NameTab: 'Rejected',
-                        BadgeTab: 9999,
-                    },
-                ]}
-            </TabsNavigation>
-            <TabsNavigation>
-                {[
-                    {
-                        NameTab: 'Live',
-                        BadgeTab: 99,
-                    },
-                    {
-                        NameTab: 'Scheduled',
-                        BadgeTab: 0,
-                    },
-                ]}
-            </TabsNavigation>
-        </div>
-    );
+  return (
+    <div className="App">
+      {/* <!-- your component goes here to test the rendering! --> */}
+      <TabsNavigation>
+        <PocketCustomTab title="Prospects" badge="3333" />
+        <PocketCustomTab title="Snoozed" badge="222" />
+        <PocketCustomTab title="Approved" badge="11" />
+        <PocketCustomTab title="Rejected" badge="0" />
+      </TabsNavigation>
+      <TabsNavigation>
+        <PocketCustomTab title="Live" badge="0" />
+        <PocketCustomTab title="Scheduled" badge="1" />
+      </TabsNavigation>
+    </div>
+  );
 }
 
 export default App;
